Create Compiler class and properly add functions.
Huge commit, didn't mean it to be this way. I created a `Compiler` class
that we can pass to builtins, and also added a `+` builtin.

I'm not really sure how to do closures or if I accidentally implemented
them already, but I'm happy with how it's going.

Now I just need to add some proper builtins and var-arg support.